Phase 2 Complete: Fix type system propagation across all modules
- Updated ui.rs: Changed HashMap keys from i64 to u32, fixed address arithmetic
- Updated linter.rs: Removed RV64 instruction variants, converted address types
- Updated execution.rs: Fixed linter initialization with u32 addresses
- Updated elf.rs: Converted address and symbol handling from i64 to u32
- Fixed all register/address type mismatches throughout the codebase

Compilation now succeeds with no errors. Project is ready for Phase 3 (ELF loader
updates for RV32 support and test suite conversion).

+++ src/execution.rs
@@ -334,7 +334,7 @@ pub fn trace(
     max_steps: usize,
     mode: &str,
 ) -> Vec<Effects> {
-    let mut linter = Linter::new(m.get_reg(2));
+    let mut linter = Linter::new(m.get_reg(2) as u32);
     let mut sequence: Vec<Effects> = Vec::new();
     let mut i = 0;
     let echo_in = [/* "run", */ "debug"].contains(&mode) && !io::stdin().is_tty();
